PR feedback - more expressive name for data provider
